+++ src/eca/config.clj
@@ -119,7 +119,7 @@
    :index {:ignoreFiles [{:type :gitignore}]
            :repoMap {:maxTotalEntries 800
                      :maxEntriesPerDir 50}}
-   :completion {:model "openai/gpt-4o"
+   :completion {:model "openai/gpt-4.1"
                 :systemPromptFile "prompts/inline_completion.md"}})