+++ src/eca/features/completion.clj
@@ -62,14 +62,20 @@
         {:keys [line character]} position
         input-code (insert-completion-tag doc-text position)
         instructions (f.prompt/inline-completion-prompt config)
-        {:keys [error result]} (llm-api/sync-prompt!
-                                {:provider provider
-                                 :model model
-                                 :config config
-                                 :prompt input-code
-                                 :instructions instructions
-                                 :provider-auth provider-auth
-                                 :model-capabilities model-capabilities})]
+        {:keys [error result]} (deref (future
+                                        (llm-api/sync-prompt!
+                                         {:provider provider
+                                          :model model
+                                          :config config
+                                          :prompt input-code
+                                          :instructions instructions
+                                          :provider-auth provider-auth
+                                          :model-capabilities (assoc model-capabilities
+                                                                     :reason? false
+                                                                     :tools false
+                                                                     :web-search false)}))
+                                      30000 ;; TODO move to config
+                                      {:error {:message "Timeout waiting for completion"}})]
     (cond
       (:message error)
       {:error {:type :warning